keyboard: move KeyboardState to dedicated module
