+++ src/ei/ei_ifs/ei_seat.rs
@@ -17,6 +17,7 @@ use {
         },
         fixed::Fixed,
         ifs::wl_seat::{wl_pointer::PendingScroll, WlSeatGlobal},
+        keyboard::{DynKeyboardState, KeyboardState, KeyboardStateId},
         leaks::Tracker,
         tree::Node,
         utils::{array, bitflags::BitflagsExt, clonecell::CloneCell},
@@ -26,7 +27,6 @@ use {
             },
             EiSeatId,
         },
-        xkbcommon::{DynKeyboardState, KeyboardState, KeyboardStateId},
     },
     std::{cell::Cell, rc::Rc},
     thiserror::Error,
