+++ src/ifs/wl_seat.rs
@@ -72,6 +72,7 @@ use {
             },
             xdg_toplevel_drag_v1::XdgToplevelDragV1,
         },
+        keyboard::{DynKeyboardState, KeyboardState},
         leaks::Tracker,
         object::{Object, Version},
         rect::Rect,
@@ -91,7 +92,7 @@ use {
             ZwpTextInputV3Id,
         },
         wire_ei::EiSeatId,
-        xkbcommon::{DynKeyboardState, KeyboardState, KeymapId, XkbKeymap, XkbState},
+        xkbcommon::{KeymapId, XkbKeymap, XkbState},
     },
     ahash::AHashMap,
     smallvec::SmallVec,
