+++ src/ifs/wl_seat/event_handling.rs
@@ -33,13 +33,14 @@ use {
             },
             wl_surface::{xdg_surface::xdg_popup::XdgPopup, WlSurface},
         },
+        keyboard::KeyboardState,
         object::Version,
         rect::Rect,
         state::DeviceHandlerData,
         tree::{Direction, Node, ToplevelNode},
         utils::{bitflags::BitflagsExt, hash_map_ext::HashMapExt, smallmap::SmallMap},
         wire::WlDataOfferId,
-        xkbcommon::{KeyboardState, XkbState, XKB_KEY_DOWN, XKB_KEY_UP},
+        xkbcommon::{XkbState, XKB_KEY_DOWN, XKB_KEY_UP},
     },
     isnt::std_1::primitive::{IsntSlice2Ext, IsntSliceExt},
     jay_config::{
