+++ src/ifs/wl_seat/zwp_virtual_keyboard_manager_v1.rs
@@ -3,10 +3,10 @@ use {
         client::{Client, ClientCaps, ClientError, CAP_VIRTUAL_KEYBOARD_MANAGER},
         globals::{Global, GlobalName},
         ifs::wl_seat::zwp_virtual_keyboard_v1::ZwpVirtualKeyboardV1,
+        keyboard::KeyboardState,
         leaks::Tracker,
         object::{Object, Version},
         wire::{zwp_virtual_keyboard_manager_v1::*, ZwpVirtualKeyboardManagerV1Id},
-        xkbcommon::KeyboardState,
     },
     std::{cell::RefCell, rc::Rc},
     thiserror::Error,
