+++ src/ifs/wl_seat/zwp_virtual_keyboard_v1.rs
@@ -9,10 +9,11 @@ use {
             },
             wl_surface::WlSurface,
         },
+        keyboard::KeyboardState,
         leaks::Tracker,
         object::{Object, Version},
         wire::{zwp_virtual_keyboard_v1::*, ZwpVirtualKeyboardV1Id},
-        xkbcommon::{KeyboardState, XkbCommonError},
+        xkbcommon::XkbCommonError,
     },
     std::{cell::RefCell, rc::Rc},
     thiserror::Error,
