+++ src/ifs/wl_surface.rs
@@ -65,6 +65,7 @@ use {
             zwp_linux_dmabuf_feedback_v1::ZwpLinuxDmabufFeedbackV1,
         },
         io_uring::IoUringError,
+        keyboard::KeyboardState,
         leaks::Tracker,
         object::{Object, Version},
         rect::{DamageQueue, Rect, Region},
@@ -88,7 +89,6 @@ use {
             wl_surface::*, WlOutputId, WlSurfaceId, ZwpIdleInhibitorV1Id,
             ZwpLinuxDmabufFeedbackV1Id,
         },
-        xkbcommon::KeyboardState,
         xwayland::XWaylandEvent,
     },
     ahash::AHashMap,
