+++ src/state.rs
@@ -66,6 +66,7 @@ use {
             zwp_linux_dmabuf_v1::ZwpLinuxDmabufV1Global,
         },
         io_uring::IoUring,
+        keyboard::KeyboardStateIds,
         leaks::Tracker,
         logger::Logger,
         rect::{Rect, Region},
@@ -99,7 +100,7 @@ use {
             ExtForeignToplevelListV1Id, ExtIdleNotificationV1Id, JayRenderCtxId, JaySeatEventsId,
             JayWorkspaceWatcherId, ZwpLinuxDmabufFeedbackV1Id,
         },
-        xkbcommon::{KeyboardStateIds, XkbContext, XkbKeymap, XkbState},
+        xkbcommon::{XkbContext, XkbKeymap, XkbState},
         xwayland::{self, XWaylandEvent},
     },
     ahash::{AHashMap, AHashSet},
